feat(console): separate health-check Dashboard in dedicated page
