+++ gravitee-apim-console-webui/src/management/api/proxy/health-check-dashboard/_healthcheck.scss
