+++ gravitee-apim-console-webui/src/management/api/proxy/health-check-dashboard/healthcheck-log.controller.ts
@@ -35,7 +35,7 @@ class ApiHealthCheckLogController {
 
   backToHealthcheck() {
     const query = JSON.parse(this.$window.localStorage.lastHealthCheckQuery);
-    this.$state.go('management.apis.detail.proxy.healthcheck.visualize', {
+    this.$state.go('management.apis.detail.proxy.healthCheckDashboard.visualize', {
       page: query.page,
       size: query.size,
       from: query.from,
