+++ gravitee-apim-console-webui/src/management/api/proxy/health-check-dashboard/healthcheck-visualize.html
@@ -18,7 +18,7 @@
 <div class="gv-forms gv-forms-fluid" layout="column">
   <div class="gv-forms-header">
     <h1>Health-check</h1>
-    <a permission permission-only="'api-health-u'" ui-sref="management.apis.detail.proxy.healthcheck.configure">Configure Health-check</a>
+    <a permission permission-only="'api-health-u'" ui-sref="management.apis.detail.proxy.healthcheck">Configure Health-check</a>
   </div>
   <div class="gv-form">
     <h2></h2>
@@ -297,7 +297,7 @@
                                 ng-if="!log.success"
                                 icon="remove_red_eye"
                                 size="16px"
-                                ui-sref="management.apis.detail.proxy.healthcheck.log({log: log.id})"
+                                ui-sref="management.apis.detail.proxy.healthCheckDashboard.log({log: log.id})"
                               ></ng-md-icon>
                             </td>
                           </tr>
