+++ gravitee-apim-console-webui/src/management/api/proxy/health-check-dashboard/healthcheck.html
