+++ gravitee-apim-console-webui/src/management/api/proxy/health-check/api-proxy-health-check.component.scss
@@ -10,3 +10,8 @@ $typography: map.get(gio.$mat-theme, typography);
 :host {
   @include gio-layout.gio-responsive-margin-container;
 }
+
+
+.go-to-dashboard {
+  float: right;
+}
