+++ gravitee-apim-console-webui/src/management/dashboard/apis-status-dashboard/apis-status-dashboard.html
@@ -76,7 +76,7 @@
             </md-card-avatar>
             <md-card-header-text>
               <span class="md-title" title="{{api.name}} {{api.version}}">{{api.name}} <small>({{api.version}})</small></span>
-              <a ui-sref="management.apis.detail.proxy.healthcheck.visualize({apiId: api.id})">View detail</a>
+              <a ui-sref="management.apis.detail.proxy.healthCheckDashboard.visualize({apiId: api.id})">View detail</a>
             </md-card-header-text>
             <md-card-avatar ng-if="$ctrl.isInError(api) || $ctrl.isInWarning(api)">
               <ng-md-icon icon="error" size="40px"></ng-md-icon>
